Add additional compiler checks and fix resultant warnings,

+++ libhobbes/hobbes_cmd_queue.h
@@ -19,7 +19,7 @@ extern "C" {
 #include "xemem.h"
 #include "hobbes_cmds.h"
 
-#define HCQ_INVALID_CMD    (-1)
+#define HCQ_INVALID_CMD    ((uint64_t)-1)
 #define HCQ_INVALID_HANDLE (NULL)
 
 typedef enum {
